+++ lib/features/user_action_pe/list/user_action_pe_list_actions.dart
@@ -1,4 +1,4 @@
-import 'package:pass_emploi_app/models/user_action_PE.dart';
+import 'package:pass_emploi_app/models/user_action_pe.dart';
 
 class UserActionPEListRequestAction {}
