+++ lib/widgets/cards/user_action_pe_card.dart
@@ -1,19 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:pass_emploi_app/presentation/user_action_PE/user_action_PE_view_model.dart';
+import 'package:pass_emploi_app/presentation/user_action_pe/user_action_pe_view_model.dart';
 import 'package:pass_emploi_app/ui/app_colors.dart';
 import 'package:pass_emploi_app/ui/margins.dart';
 import 'package:pass_emploi_app/ui/shadows.dart';
 import 'package:pass_emploi_app/ui/strings.dart';
 import 'package:pass_emploi_app/ui/text_styles.dart';
-import 'package:pass_emploi_app/widgets/tags/status_tag_PE.dart';
-
-import '../../ui/drawables.dart';
+import 'package:pass_emploi_app/widgets/tags/status_tag_pe.dart';
+import 'package:pass_emploi_app/ui/drawables.dart';
 
 class UserActionPECard extends StatelessWidget {
   final UserActionPEViewModel viewModel;
 
-  const UserActionPECard({Key? key, required this.viewModel}) : super(key: key);
+  const UserActionPECard({required this.viewModel}) : super();
 
   @override
   Widget build(BuildContext context) {
